Add event listeners

+++ nikola/src/entity/entity.cpp
@@ -36,7 +36,7 @@ EntityID entity_world_create_entity(EntityWorld& world,
 
   Event event = {
     .type    = EVENT_ENTITY_ADDED, 
-    .entt_id = (u32)entt,
+    .entt_id = entt,
   };
   event_dispatch(event);
 
@@ -49,7 +49,7 @@ void entity_world_destroy_entity(EntityWorld& world, EntityID& entt) {
 
   Event event = {
     .type    = EVENT_ENTITY_DESTROYED, 
-    .entt_id = (u32)entt,
+    .entt_id = entt,
   };
   event_dispatch(event);
